Conflicts:
	README.md
	finagle-core/src/main/scala/com/twitter/finagle/builder/ServerBuilder.scala
	finagle-exception/src/test/scala/com/twitter/finagle/exception/ReporterSpec.scala
	sbt

+++ finagle-memcached/src/test/scala/com/twitter/finagle/memcached/integration/ClientSpec.scala
@@ -4,13 +4,13 @@ import _root_.java.io.ByteArrayOutputStream
 import _root_.java.lang.{Boolean => JBoolean}
 import java.net.{SocketAddress, InetSocketAddress}
 import com.twitter.common.application.ShutdownRegistry.ShutdownRegistryImpl
-import com.twitter.common.zookeeper.ServerSet.EndpointStatus
 import com.twitter.common.zookeeper.{ZooKeeperUtils, ServerSets, ZooKeeperClient}
+import com.twitter.common.zookeeper.ServerSet.EndpointStatus
 import com.twitter.common.zookeeper.testing.ZooKeeperTestServer
 import com.twitter.concurrent.Spool
 import com.twitter.concurrent.Spool.*::
 import com.twitter.conversions.time._
-import com.twitter.finagle.Group
+import com.twitter.finagle.{Group, MemcachedClient, Name, WriteException}
 import com.twitter.finagle.builder.{Cluster, ClientBuilder}
 import com.twitter.finagle.memcached.{CacheNode, CacheNodeGroup, CachePoolCluster, CachePoolConfig,
   Client, KetamaClientBuilder, PartitionedClient}
@@ -20,7 +20,6 @@ import com.twitter.finagle.memcached.replication._
 import com.twitter.finagle.memcached.util.ChannelBufferUtils._
 import com.twitter.finagle.stats.SummarizingStatsReceiver
 import com.twitter.finagle.zookeeper.ZookeeperServerSetCluster
-import com.twitter.finagle.{MemcachedClient, WriteException}
 import com.twitter.util.{Await, Duration, Future, Return, Throw}
 import org.jboss.netty.util.CharsetUtil
 import org.specs.SpecificationWithJUnit
@@ -200,6 +199,16 @@ class ClientSpec extends SpecificationWithJUnit {
         val client = KetamaClientBuilder()
           .nodes("localhost:%d,localhost:%d".format(address1.getPort, address2.getPort))
           .build()
+
+        Await.result(client.delete("foo"))
+        Await.result(client.get("foo")) mustEqual None
+        Await.result(client.set("foo", "bar"))
+        Await.result(client.get("foo")).get.toString(CharsetUtil.UTF_8) mustEqual "bar"
+      }
+
+      "using Name doesn't blow up" in {
+        val name = Name.bound(address1, address2)
+        val client = KetamaClientBuilder().dest(name).build()
 
         Await.result(client.delete("foo"))
         Await.result(client.get("foo")) mustEqual None
@@ -755,19 +764,36 @@ class ClientSpec extends SpecificationWithJUnit {
       testServers = List()
     }
 
-    "with static servers list" in {
-      val client = MemcachedClient.newKetamaClient(
-        group = "twcache!localhost:%d,localhost:%d".format(testServers(0).address.getPort, testServers(1).address.getPort))
+    if (!Option(System.getProperty("SKIP_FLAKY")).isDefined) {
+      "with unmanaged regular zk serverset" in {
+        val client = MemcachedClient.newKetamaClient(
+          "zk!localhost:"+zookeeperServerPort+"!"+zkPath).asInstanceOf[PartitionedClient]
 
-      Await.result(client.delete("foo"))
-      Await.result(client.get("foo")) mustEqual None
-      Await.result(client.set("foo", "bar"))
-      Await.result(client.get("foo")).get.toString(CharsetUtil.UTF_8) mustEqual "bar"
+        // Wait for group to contain members
+        Thread.sleep(5000)
+
+        val count = 100
+        (0 until count).foreach{
+          n => {
+            client.set("foo"+n, "bar"+n)()
+          }
+        }
+
+        (0 until count).foreach {
+          n => {
+            val c = client.clientOf("foo"+n)
+            c.get("foo"+n)().get.toString(CharsetUtil.UTF_8) mustEqual "bar"+n
+          }
+        }
+      }
     }
 
-    "with managed cache pool" in {
+    if (!Option(System.getProperty("SKIP_FLAKY")).isDefined) "with managed cache pool" in {
       val client = MemcachedClient.newKetamaClient(
-        group = "twcache!localhost:"+zookeeperServerPort+"!"+zkPath).asInstanceOf[PartitionedClient]
+        "twcache!localhost:"+zookeeperServerPort+"!"+zkPath).asInstanceOf[PartitionedClient]
+
+      // Wait for group to contain members
+      Thread.sleep(5000)
 
       client.delete("foo")()
       client.get("foo")() mustEqual None
@@ -789,26 +815,14 @@ class ClientSpec extends SpecificationWithJUnit {
       }
     }
 
-    "with unmanaged regular zk serverset" in {
+    "with static servers list" in {
       val client = MemcachedClient.newKetamaClient(
-        group = "zk!localhost:"+zookeeperServerPort+"!"+zkPath).asInstanceOf[PartitionedClient]
+        "twcache!localhost:%d,localhost:%d".format(testServers(0).address.getPort, testServers(1).address.getPort))
 
-      // Wait for group to contain members
-      Thread.sleep(5000)
-
-      val count = 100
-      (0 until count).foreach{
-        n => {
-          client.set("foo"+n, "bar"+n)()
-        }
-      }
-
-      (0 until count).foreach {
-        n => {
-          val c = client.clientOf("foo"+n)
-          c.get("foo"+n)().get.toString(CharsetUtil.UTF_8) mustEqual "bar"+n
-        }
-      }
+      Await.result(client.delete("foo"))
+      Await.result(client.get("foo")) mustEqual None
+      Await.result(client.set("foo", "bar"))
+      Await.result(client.get("foo")).get.toString(CharsetUtil.UTF_8) mustEqual "bar"
     }
   }
 }
